requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.7
django_redis==5.2.0
django_filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
easy_thumbnails==2.8.5
idna==3.4
kombu==5.2.4
mailjet_rest==1.3.4
prompt-toolkit==3.0.38
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
